Credit-Transfer-20260428/4-Computational Visual Media/Assignments/Assignment1/22214624-Shanshan Zhong/code/requirements.txt
matplotlib==3.3.4
numpy==1.19.5
Pillow==9.3.0
progress==1.6
torch==1.10.1
torchvision==0.11.2
